feat:新增回傳成功或失敗的功能 新增一種MessageType，UndefinedType 增加移除User時同時會移除房間和列隊的功能

+++ uptalk_spring/src/main/java/com/UPT/uptalk_spring/service/impl/SessionPoolServiceImpl.java
@@ -1,17 +1,20 @@
 package com.UPT.uptalk_spring.service.impl;
 
-import com.UPT.uptalk_spring.model.ExistRoom;
-import com.UPT.uptalk_spring.model.MessageMetaData;
-import com.UPT.uptalk_spring.model.User;
-import com.UPT.uptalk_spring.model.UserInfo;
+import com.UPT.uptalk_spring.model.*;
 import com.UPT.uptalk_spring.repository.IExistRoomRepository;
 import com.UPT.uptalk_spring.service.IRoomService;
 import com.UPT.uptalk_spring.service.ISessionPoolService;
-import lombok.NoArgsConstructor;
+import com.UPT.uptalk_spring.utils.websocket.MessageMetaDataUtil;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import javax.websocket.Session;
+import java.io.IOException;
 import java.util.*;
+
+import static com.UPT.uptalk_spring.model.MessageType.CreateRoom;
+import static com.UPT.uptalk_spring.model.MessageType.JoinRoom;
 
 /**
  * @Title: SessionPoolServiceImpl
@@ -20,8 +23,10 @@ import java.util.*;
  * @time: 2022/5/13
  */
 @Slf4j
-@NoArgsConstructor
+@Service
 public class SessionPoolServiceImpl implements ISessionPoolService {
+
+    @Autowired
     private IExistRoomRepository IExistRoomRepository;
     private static Map<Session, User> userPool = new HashMap<Session, User>();
     private static Map<Integer, RoomServiceImpl> roomPool = new HashMap<Integer, RoomServiceImpl>();
@@ -48,6 +53,8 @@ public class SessionPoolServiceImpl implements ISessionPoolService {
      */
     @Override
     public void removeUser(Session session) {
+        leaveQueue(session);
+        quitRoom(session);
         userPool.remove(session);
         log.info(session.getId() + ":連線退出，目前總連線數:" + userPool.size());
     }
@@ -59,17 +66,20 @@ public class SessionPoolServiceImpl implements ISessionPoolService {
      * @return 建立出的房間Id
      */
     public Integer createRoom(Session session) {
-        RoomServiceImpl roomServiceImpl = new RoomServiceImpl();
-        User user = userPool.get(session);
-        int roomId = generateRoomId();
+        try {
+            User user = userPool.get(session);
+            int roomId = generateRoomId();
+            RoomServiceImpl roomServiceImpl = new RoomServiceImpl();
+            roomServiceImpl.setId(roomId);
+            roomServiceImpl.addUser(session, user);
 
-        roomServiceImpl.setId(roomId);
-        roomServiceImpl.addUser(session, user);
+            roomPool.put(roomId, roomServiceImpl);
 
-        roomPool.put(roomId, roomServiceImpl);
-
-        log.info(session.getId() + "建立房間:" + roomId + " 目前總房間數:" + roomPool.size());
-        return roomId;
+            log.info(session.getId() + ":建立房間:" + roomId + " 目前總房間數:" + roomPool.size());
+            return roomId;
+        } catch (Exception e) {
+            return null;
+        }
     }
 
     /**
@@ -77,24 +87,41 @@ public class SessionPoolServiceImpl implements ISessionPoolService {
      *
      * @param session 要加入房間的session
      * @param roomId  要加入的房間Id
+     * @return 是否成功
      */
-    public void joinRoom(Session session, int roomId) {
-        User user = userPool.get(session);
-        RoomServiceImpl roomServiceImpl = roomPool.get(roomId);
-        roomServiceImpl.addUser(session, user);
+    public Boolean joinRoom(Session session, int roomId) {
+        try {
+            User user = userPool.get(session);
+            RoomServiceImpl roomServiceImpl = roomPool.get(roomId);
+            roomServiceImpl.addUser(session, user);
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
     }
 
     /**
      * 離開房間
      *
      * @param session 要離開房間的session
+     * @return 是否成功
      */
-    public void quitRoom(Session session) {
-        User user = userPool.get(session);
-        IRoomService roomService = user.getRoomService();
-        roomService.removeUser(session, user);
-        if (roomService.getUserPool().size() == 0) {
-            roomPool.remove(roomService.getId());
+    public Boolean quitRoom(Session session) {
+        try {
+            User user = userPool.get(session);
+            IRoomService roomService = user.getRoomService();
+            if (roomService != null) {
+                roomService.removeUser(session, user);
+                if (roomService.getUserPool().size() == 0) {
+                    int roomId = roomService.getId();
+                    IExistRoomRepository.deleteRoomByRoomNumber(roomId);
+                    roomPool.remove(roomId);
+                    log.info(session.getId() + ":移除房間:" + roomId + " 目前總房間數:" + roomPool.size());
+                }
+            }
+            return true;
+        } catch (Exception e) {
+            return false;
         }
     }
 
@@ -103,35 +130,61 @@ public class SessionPoolServiceImpl implements ISessionPoolService {
      *
      * @param session         要送出訊息的session
      * @param messageMetaData 要送出的訊息物件
+     * @return 是否成功
      */
-    public void sendMessage(Session session, MessageMetaData messageMetaData) {
-        User user = userPool.get(session);
-        IRoomService roomService = user.getRoomService();
-        roomService.sendMessage(session, messageMetaData);
+    public Boolean sendMessage(Session session, MessageMetaData messageMetaData) {
+        try {
+            User user = userPool.get(session);
+            IRoomService roomService = user.getRoomService();
+            roomService.sendMessage(session, messageMetaData);
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
     }
 
     /**
      * 加入列隊
      *
      * @param session 要加入列隊的session
+     * @return 是否成功
      */
-    public void joinQueue(Session session) {
-        queuePool.add(session);
-        if (queuePool.size() >= 2) {
-            int roomId = this.createRoom(queuePool.poll());
-            this.joinRoom(queuePool.poll(), roomId);
+    public Boolean joinQueue(Session session) {
+        try {
+            queuePool.add(session);
+            log.info(session.getId() + ":加入列隊 目前列隊人數:" + queuePool.size());
+            if (queuePool.size() >= 2) {
+                Session session1 = queuePool.poll();
+                Session session2 = queuePool.poll();
+                Integer roomId = this.createRoom(session1);
+                if (roomId != null) {
+                    sendCommand(session1, CreateRoom, "success");
+                }
+                if (this.joinRoom(session2, roomId)) {
+                    sendCommand(session2, JoinRoom, "success");
+                }
+                log.info(session.getId() + ":列隊人數超過兩人，創建房間 目前列隊人數:" + queuePool.size());
+            }
+            return true;
+        } catch (Exception e) {
+            return false;
         }
-        log.info(session.getId() + "加入列隊 目前列隊人數:" + queuePool.size());
     }
 
     /**
      * 離開列隊
      *
      * @param session 要離開列隊的session
+     * @return 是否成功
      */
-    public void leaveQueue(Session session) {
-        queuePool.remove(session);
-        log.info(session.getId() + "離開列隊 目前列隊人數:" + queuePool.size());
+    public Boolean leaveQueue(Session session) {
+        try {
+            queuePool.remove(session);
+            log.info(session.getId() + ":離開列隊 目前列隊人數:" + queuePool.size());
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
     }
 
     /**
@@ -141,11 +194,24 @@ public class SessionPoolServiceImpl implements ISessionPoolService {
      */
     public int generateRoomId() {
         Optional<ExistRoom> existRoom;
+        int roomId;
         do {
-            existRoom = this.IExistRoomRepository.findExistRoomByRoomNumber((int) (Math.random() * 100000));
+            roomId = (int) (Math.random() * 100000)+100000;
+            existRoom = IExistRoomRepository.findExistRoomByRoomNumber(roomId);
         } while (existRoom.isPresent());
-        int roomId = existRoom.get().getRoomNumber();
+        IExistRoomRepository.save(new ExistRoom(roomId));
         log.info("產生房號:" + roomId);
         return roomId;
+    }
+
+    public void sendCommand(Session session, MessageType type, String command) {
+        MessageMetaData messageMetaData = new MessageMetaData(type, command);
+        String returnMessage = MessageMetaDataUtil.encodeMessageMetaData(messageMetaData);
+        try {
+            session.getBasicRemote().sendText(returnMessage);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        log.info(session.getId() + "送出命令:" + messageMetaData);
     }
 }
